+++ core/os/device/bind/BUILD.bazel
@@ -20,12 +20,13 @@ go_library(
     name = "go_default_library",
     srcs = [
         "bind.go",
+        "bind_posix.go",
+        "bind_windows.go",
+        "desktop.go",
         "device.go",
         "doc.go",
         "registry.go",
         "simple.go",
-        "simple_posix.go",
-        "simple_windows.go",
     ],
     embed = [":bind_go_proto"],
     importpath = "github.com/google/gapid/core/os/device/bind",
